Finishes transition from InfluxQL to Flux 	modified:   Syslog-Dashboard.json
Changes the bucket to v.defaultBucket for both panels.

For the syslog count panel, adds the legend to the time series,
fixes the names of each type, and fixes the colors to match the table.

For the Syslog Messages panel, restores the message search and
collapses it to one table